refactor(convex): use verifyChatOwnership for messages functions only
